Extract the classification reference behind the seam, not through it
The size ratchet in services/api/test_file_sizes.py refused the previous
commit's growth of apps/web/src/api/client.ts (554 -> 562), which is the pin
working exactly as its comment says it should. The answer is the one the file
already documents: extract, never raise the cap. `disciplineTree` came across
with the new accessor rather than being left behind -- they are one question,
*what does the served vocabulary say?*, answered from one cached request, and
splitting them would put the cache in a different file from its readers.
client.ts 554 -> 540.

The extraction left two residues, and both were caught by gates rather than by
me re-reading the diff:

  * DOC-STRAND (apps/web/src/tooling/docStranded.test.ts, and its api/ sibling)
    found the doc comment whose declaration had moved out from under it --
    stranded directly above the NEXT method's comment, where it would be read
    as documenting the 5D heatmap. Moved to the alias it actually describes.
  * eslint found the now-unused `DisciplineTree` import.

A gate that only fires on the day someone does the refactor is doing the job
nobody can do by eye.

Web suite 2433/2433; lint, typecheck, build clean; ruff clean over the whole
tree; test_file_sizes and test_claude_md_gates pass.
